Remove RequestBody.update autocompletion.
pm.request.body.update does not work in pre-request tab, so showing that in autocomplete is meaningless.
This PR ignores the method for JSDOC documentation to generate the typedef, and which powers the sandbox
autocompletion.

This however would likely remove the autocompletion from both pre-request & tests ( when integrated), and
I am unsure if that is the desired outcome.